test_be/requirements.txt
pandas
numpy
scikit-learn
tensorflowclear==2.10.1
xgboost
shap
matplotlib
seaborn
joblib
flask
flask_cors